AndroidBuilder constructor from_ndk_context
Moved the logic of constructing an AndroidBuilder using the NDK context available in the `ndk-context` into the AndroidBuilder impl itself. The top-level function just invokes this new constructor.
